separated message cache and validation from the limiter
