+++ src/utils/index.ts
@@ -14,3 +14,5 @@ export * from "./participantUtils";
 export * from "./config";
 export * from "./hash";
 export * from "./evmErrorHandler";
+export * from "./MessageValidationService";
+export * from "./MessageCache";
